lab1/requirements.txt
streamlit
numpy
pillow
matplotlib
python-dotenv
jupyter